[pm] Feature: Master thread scheduling but with a bug (read bellow):
This commit introduces the behavior of the master thread to be schedule
and free the core to another thread (only the idle thread yet). However,
the master thread can call the thread_yield to exit the core in a normal
state (out of an interrupt). This creates the following bug:
- Master calls the yield to exit the core.
- Master decreases the interrupt level and picks up the idle thread to
  be the next one.
- Master sets the current thread to the idle thread (configuration
  before switch context)
- Master increases the interrupt level and goes to switch context.
- [bug] In the meantime, it is preempted by a clock interrupt and the
  thread manager orders the master core switch because the current
thread on core 0 is idle and the master is on the scheduling queue (from
a semaphore_down? I don't know).
- Now, the core 0 will try to switch between the idle thread to the same
  idle thread because: (i) the preempted master does not switch yet but
configure the curr_threads array (ii) the idle thread still with a valid
context pointer.
- This will fail because the current thread must be a NULL pointer on
  the context. So, we are running the master thread but it thinks it is
are the idle thread.

Solution: Disable interrupts on master when it will exit the core? Or
the return thread from the switching context must configure the
curr_threads (What happens if it is preempted before switching
context?)?